+++ NEWS.md
@@ -1,3 +1,12 @@
+# RSuite 0.36 (2019-03-06)
+  * #184: detecting declaration of dependencies in scripts with "library (...)".
+  * handling of branches in Git RC adapter fixed.
+  * #183: local Git repository is created while starting project if not RC detected
+  * when checking for changes in Git repository only changes in project folder are
+    considered.
+  * prj_build have new parameter tag which enforces tagging packages with RC revision
+    before build.
+  
 # RSuite 0.35 (2019-01-07)
   * Fix in dependecy resolving process. Upper requirements enforse subdependencies
     version selection not to present version collisions.